+++ src/defaults.ts
@@ -1,3 +1,5 @@
+import { join } from "path";
+
 import type { Options } from "./types.js";
 
 const EXTENSIONS = [
@@ -8,6 +10,27 @@ const EXTENSIONS = [
 ];
 
 const DEFAULT_FILES = EXTENSIONS.map((ext) => `**/*.${ext}`);
+
+const ensureExtensions = (inputPath: string, extensions = EXTENSIONS): string[] => {
+    if (inputPath.includes("*")) {
+        return [inputPath];
+    }
+
+    const segments = inputPath.split("/").filter(Boolean);
+    if (!segments.length) {
+        // weird path, let's just return it as is
+        return [inputPath];
+    }
+
+    const lastSegment = segments[segments.length - 1];
+    if (lastSegment?.includes(".")) {
+        return [inputPath];
+    }
+
+    return extensions.map((ext) => {
+        return join(inputPath, "**", `*.${ext}`);
+    });
+};
 
 const DEFAULT_INDENT = 4;
 
@@ -36,5 +59,6 @@ const defaultOptions: Options & Pick<Required<Options>, "files" | "config"> = {
 
 export {
     defaultOptions,
+    ensureExtensions,
     EXTENSIONS,
 };